webkit/data/layout_tests/platform/chromium-win/LayoutTests/fast/dom/java-applet-calls-expected.txt

Requires Java plugin installed and enabled. This tests that calling functions on Java Applet works.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



PASS document.testapplet.getString() == 'abcdefg' is true
PASS successfullyParsed is true

TEST COMPLETE

